# services/graph_harness_api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HarnessProposalDecisionRequest": {
        "description": "Promote or reject one proposal.",
        "properties": {
          "metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Input"
            },
            "title": "Metadata",
            "type": "object"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "HarnessProposalDecisionRequest",
        "type": "object"
      },
      "HarnessProposalListResponse": {
        "description": "List response for proposals.",
        "properties": {
          "proposals": {
            "items": {
              "$ref": "#/components/schemas/HarnessProposalResponse"
            },
            "title": "Proposals",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "proposals",
          "total"
        ],
        "title": "HarnessProposalListResponse",
        "type": "object"
      },
      "HarnessProposalResponse": {
        "description": "Serialized proposal record.",
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "evidence_bundle": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JSONValue-Output"
              },
              "type": "object"
            },
            "title": "Evidence Bundle",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "title": "Metadata",
            "type": "object"
          },
          "payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "title": "Payload",
            "type": "object"
          },
          "proposal_type": {
            "title": "Proposal Type",
            "type": "string"
          },
          "ranking_score": {
            "title": "Ranking Score",
            "type": "number"
          },
          "reasoning_path": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "title": "Reasoning Path",
            "type": "object"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source_key": {
            "title": "Source Key",
            "type": "string"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "space_id": {
            "title": "Space Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "space_id",
          "run_id",
          "proposal_type",
          "source_kind",
          "source_key",
          "title",
          "summary",
          "status",
          "confidence",
          "ranking_score",
          "reasoning_path",
          "evidence_bundle",
          "payload",
          "metadata",
          "decision_reason",
          "decided_at",
          "created_at",
          "updated_at"
        ],
        "title": "HarnessProposalResponse",
        "type": "object"
      },
...
      "JSONValue-Input": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Input"
            },
            "type": "object"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JSONValue-Input"
            },
            "type": "array"
          },
          {
            "type": "null"
          }
        ]
      },
      "JSONValue-Output": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "type": "object"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "type": "array"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MED13 Graph Harness API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/spaces/{space_id}/proposals": {
      "get": {
        "description": "Return proposals for one research space.",
        "operationId": "list_proposals_v1_spaces__space_id__proposals_get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "proposal_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Proposal Type"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessProposalListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List staged proposals",
        "tags": [
          "proposals"
        ]
      }
    },
    "/v1/spaces/{space_id}/proposals/{proposal_id}": {
      "get": {
        "description": "Return one proposal with evidence and ranking.",
        "operationId": "get_proposal_v1_spaces__space_id__proposals__proposal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get one staged proposal",
        "tags": [
          "proposals"
        ]
      }
    },
    "/v1/spaces/{space_id}/proposals/{proposal_id}/promote": {
      "post": {
        "description": "Promote one proposal into the reviewed state.",
        "operationId": "promote_proposal_v1_spaces__space_id__proposals__proposal_id__promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessProposalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Promote one reviewed proposal",
        "tags": [
          "proposals"
        ]
      }
    },
    "/v1/spaces/{space_id}/proposals/{proposal_id}/reject": {
      "post": {
        "description": "Reject one proposal without touching the graph ledger.",
        "operationId": "reject_proposal_v1_spaces__space_id__proposals__proposal_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessProposalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject one staged proposal",
        "tags": [
          "proposals"
        ]
      }
    },
...
  }
}